+++ packages/shared/src/ddb/index.ts
@@ -1,3 +1,3 @@
-export { getDdbClient, isConditionalCheckFailed } from "./client";
+export { getDdbClient, isConditionalCheckFailed, __resetDdbClientForTests } from "./client";
 export { encodeNextToken, decodeNextToken, type PaginatedResult } from "./pagination";
 export { BaseRepository, type BaseRepositoryOptions, type QueryOptions } from "./base-repository";